fixing many bugs after sap server was changed. now before wan cin to inbound IIS

+++ HashtilERP/Server/DBTestVol1/Oitm.cs
@@ -259,7 +259,7 @@ namespace HashtilERP.DBTestVol1
         public short? UserSign2 { get; set; }
         public string Spec { get; set; }
         public string TaxCtg { get; set; }
-        public short? Series { get; set; }
+        public int? Series { get; set; }
         public int? Number { get; set; }
         public int? FuelCode { get; set; }
         public string BeverTblC { get; set; }
